For the ProgramTests.cs file, in the GetPercentCorrect_ReturnsExpectedPercentage() method, changed the incoming parameter type from "int" to a "double". Also, in the Program.cs file, in the GetPercentCorrect() method, changed the incoming parameter from "int" to a "double".

+++ PrincessBrideTrivia/PrincessBrideTrivia.Tests/ProgramTests.cs
@@ -60,8 +60,8 @@ namespace PrincessBrideTrivia.Tests
         [DataRow(5, 10, "50%")]
         [DataRow(1, 10, "10%")]
         [DataRow(0, 10, "0%")]
-        public void GetPercentCorrect_ReturnsExpectedPercentage(int numberOfCorrectGuesses, 
-            int numberOfQuestions, string expectedString)
+        public void GetPercentCorrect_ReturnsExpectedPercentage(double numberOfCorrectGuesses, 
+            double numberOfQuestions, string expectedString)
         {
             // Arrange
